Удалено все, на что ругался спринг

+++ backend/src/main/kotlin/com/jetbrains/life_science/category/dto/CategoryDTO.kt
@@ -8,5 +8,5 @@ data class CategoryDTO(
 
     val parentId: Long?,
 
-    val order: Int
+    val order: Long
 )
